refactor: update sandboxes metrics handling and improve utility imports
- Removed the fillMetricsWithZeros function and replaced it with fillTeamMetricsWithZeros for better clarity and consistency.
- Updated import paths across various components to reflect the new utility structure.
- Cleaned up unused test files and streamlined the test suite for better maintainability.
- Enhanced the handling of metrics in the sandboxes monitoring components to improve performance and readability.
