+++ src/__test__/unit/chart-utils.test.ts
@@ -3,321 +3,11 @@ import {
   calculateYAxisMax,
   createChartSeries,
   createMonitoringChartOptions,
-  fillMetricsWithZeros,
   transformMetricsToLineData,
 } from '@/features/dashboard/sandboxes/monitoring/charts/utils'
-import type { ClientTeamMetrics } from '@/types/sandboxes.types'
 import { describe, expect, it } from 'vitest'
 
 describe('chart-utils', () => {
-  describe('fillMetricsWithZeros with overfetch handling', () => {
-    describe('Overfetch edge case - data at or beyond boundary', () => {
-      it('should NOT add zeros when last data point is beyond the end boundary', () => {
-        const data: ClientTeamMetrics = [
-          { timestamp: 1000000, concurrentSandboxes: 10, sandboxStartRate: 5 },
-          { timestamp: 1300000, concurrentSandboxes: 20, sandboxStartRate: 10 },
-          { timestamp: 1600000, concurrentSandboxes: 30, sandboxStartRate: 15 },
-          { timestamp: 2000500, concurrentSandboxes: 40, sandboxStartRate: 20 }, // beyond end
-        ]
-
-        const result = fillMetricsWithZeros(
-          data,
-          1000000,
-          2000000, // end boundary
-          300000, // 5 minute step
-          0.1
-        )
-
-        // should NOT add zeros at the end since data continues beyond boundary
-        const zerosAtEnd = result.filter(
-          (p) => p.timestamp > 2000500 && p.concurrentSandboxes === 0
-        )
-        expect(zerosAtEnd).toHaveLength(0)
-
-        // last point should be the actual data point beyond boundary
-        const lastPoint = result[result.length - 1]
-        expect(lastPoint?.timestamp).toBe(2000500)
-        expect(lastPoint?.concurrentSandboxes).toBe(40)
-      })
-
-      it('should NOT add zeros when last point is within half a step of the end', () => {
-        const step = 300000 // 5 minutes
-        const data: ClientTeamMetrics = [
-          { timestamp: 1000000, concurrentSandboxes: 10, sandboxStartRate: 5 },
-          { timestamp: 1300000, concurrentSandboxes: 20, sandboxStartRate: 10 },
-          { timestamp: 1600000, concurrentSandboxes: 30, sandboxStartRate: 15 },
-          { timestamp: 1900000, concurrentSandboxes: 40, sandboxStartRate: 20 }, // 100s from end (less than half step)
-        ]
-
-        const result = fillMetricsWithZeros(data, 1000000, 2000000, step, 0.1)
-
-        // should NOT add zeros since last point is very close to boundary
-        const zerosAtEnd = result.filter(
-          (p) => p.timestamp > 1900000 && p.concurrentSandboxes === 0
-        )
-        expect(zerosAtEnd).toHaveLength(0)
-
-        // last point should be the actual data
-        const lastPoint = result[result.length - 1]
-        expect(lastPoint?.timestamp).toBe(1900000)
-        expect(lastPoint?.concurrentSandboxes).toBe(40)
-      })
-
-      it('should add zeros when last point is more than 2 steps from end', () => {
-        const step = 300000 // 5 minutes
-        const data: ClientTeamMetrics = [
-          { timestamp: 1000000, concurrentSandboxes: 10, sandboxStartRate: 5 },
-          { timestamp: 1300000, concurrentSandboxes: 20, sandboxStartRate: 10 },
-          { timestamp: 1200000, concurrentSandboxes: 30, sandboxStartRate: 15 }, // 800s from end (> 2 steps)
-        ]
-
-        const result = fillMetricsWithZeros(data, 1000000, 2000000, step, 0.1)
-
-        // should add zeros since there's a significant gap
-        const zerosAtEnd = result.filter(
-          (p) => p.timestamp > 1300000 && p.concurrentSandboxes === 0
-        )
-        expect(zerosAtEnd.length).toBeGreaterThan(0)
-
-        // should have zero at end - 1000
-        const endZero = result.find((p) => p.timestamp === 1999000)
-        expect(endZero).toBeDefined()
-        expect(endZero?.concurrentSandboxes).toBe(0)
-      })
-    })
-
-    describe('API granularity mismatch detection', () => {
-      it('should detect and fill gaps when API returns hourly data but step is 5 minutes', () => {
-        const step = 5 * 60 * 1000 // 5 minutes
-        const data: ClientTeamMetrics = [
-          {
-            timestamp: 1609459200000,
-            concurrentSandboxes: 10,
-            sandboxStartRate: 5,
-          },
-          {
-            timestamp: 1609462800000,
-            concurrentSandboxes: 20,
-            sandboxStartRate: 10,
-          }, // 1 hour later
-          {
-            timestamp: 1609466400000,
-            concurrentSandboxes: 30,
-            sandboxStartRate: 15,
-          }, // 1 hour later
-        ]
-
-        const result = fillMetricsWithZeros(
-          data,
-          1609459200000,
-          1609470000000,
-          step,
-          0.1
-        )
-
-        // should detect the 1-hour gaps as anomalous and add zeros
-        const zerosInFirstGap = result.filter(
-          (p) =>
-            p.timestamp > 1609459200000 &&
-            p.timestamp < 1609462800000 &&
-            p.concurrentSandboxes === 0
-        )
-        expect(zerosInFirstGap.length).toBeGreaterThan(0)
-
-        // should have added zeros at expected 5-minute intervals
-        // at least one after the first point and one before the second
-        const suffixZero = zerosInFirstGap.find(
-          (p) => p.timestamp === 1609459200000 + step
-        )
-        expect(suffixZero).toBeDefined()
-
-        const prefixZero = zerosInFirstGap.find(
-          (p) => p.timestamp === 1609462800000 - step
-        )
-        expect(prefixZero).toBeDefined()
-      })
-
-      it('should handle first two points being anomalously far apart', () => {
-        const step = 5 * 60 * 1000 // 5 minutes
-        const data: ClientTeamMetrics = [
-          {
-            timestamp: 1609459200000,
-            concurrentSandboxes: 10,
-            sandboxStartRate: 5,
-          },
-          {
-            timestamp: 1609462800000,
-            concurrentSandboxes: 20,
-            sandboxStartRate: 10,
-          }, // 1 hour later (should be 5 min)
-        ]
-
-        const result = fillMetricsWithZeros(
-          data,
-          1609459200000,
-          1609466400000,
-          step,
-          0.1
-        )
-
-        // should have added zeros between the first two points
-        const zerosBetween = result.filter(
-          (p) =>
-            p.timestamp > 1609459200000 &&
-            p.timestamp < 1609462800000 &&
-            p.concurrentSandboxes === 0
-        )
-        expect(zerosBetween.length).toBeGreaterThan(0)
-
-        // verify original points are preserved
-        const originalPoints = result.filter((p) => p.concurrentSandboxes > 0)
-        expect(originalPoints).toHaveLength(2)
-      })
-    })
-
-    describe('Zoom scenarios with custom time ranges', () => {
-      it('should handle zooming into a known active region correctly', () => {
-        // simulating overfetched data (last point is from beyond the zoom range)
-        const step = 60000 // 1 minute
-        const data: ClientTeamMetrics = [
-          {
-            timestamp: 1609459200000,
-            concurrentSandboxes: 10,
-            sandboxStartRate: 5,
-          },
-          {
-            timestamp: 1609459260000,
-            concurrentSandboxes: 15,
-            sandboxStartRate: 7,
-          },
-          {
-            timestamp: 1609459320000,
-            concurrentSandboxes: 20,
-            sandboxStartRate: 10,
-          },
-          {
-            timestamp: 1609459380000,
-            concurrentSandboxes: 25,
-            sandboxStartRate: 12,
-          },
-          {
-            timestamp: 1609459440000,
-            concurrentSandboxes: 30,
-            sandboxStartRate: 15,
-          },
-          {
-            timestamp: 1609459520000,
-            concurrentSandboxes: 35,
-            sandboxStartRate: 17,
-          }, // overfetched, beyond end + tolerance
-        ]
-
-        const zoomStart = 1609459200000
-        const zoomEnd = 1609459500000 // zoom ends at 500, but we have data at 520
-
-        const result = fillMetricsWithZeros(data, zoomStart, zoomEnd, step, 0.1)
-
-        // data point at 520 is within half a step (30s) of the end (500)
-        // so it should NOT add zeros after the last data point
-        const zerosAfterLastData = result.filter(
-          (p) => p.timestamp > 1609459520000 && p.concurrentSandboxes === 0
-        )
-        expect(zerosAfterLastData).toHaveLength(0)
-
-        // should preserve all data points including overfetched ones
-        const dataPoints = result.filter((p) => p.concurrentSandboxes > 0)
-        expect(dataPoints).toHaveLength(6)
-      })
-
-      it('should add zeros when zooming into a region with gaps', () => {
-        const step = 60000 // 1 minute
-        const data: ClientTeamMetrics = [
-          {
-            timestamp: 1609459200000,
-            concurrentSandboxes: 10,
-            sandboxStartRate: 5,
-          },
-          {
-            timestamp: 1609459260000,
-            concurrentSandboxes: 15,
-            sandboxStartRate: 7,
-          },
-          // gap here - no data for several minutes
-          {
-            timestamp: 1609459500000,
-            concurrentSandboxes: 20,
-            sandboxStartRate: 10,
-          },
-          {
-            timestamp: 1609459560000,
-            concurrentSandboxes: 25,
-            sandboxStartRate: 12,
-          },
-        ]
-
-        const result = fillMetricsWithZeros(
-          data,
-          1609459200000,
-          1609459600000,
-          step,
-          0.1
-        )
-
-        // should detect the gap and add zeros
-        const zerosInGap = result.filter(
-          (p) =>
-            p.timestamp > 1609459260000 &&
-            p.timestamp < 1609459500000 &&
-            p.concurrentSandboxes === 0
-        )
-        expect(zerosInGap.length).toBeGreaterThan(0)
-      })
-    })
-
-    describe('Step tolerance behavior', () => {
-      it('should respect tolerance when detecting anomalous gaps', () => {
-        const step = 100000 // 100 seconds
-        const data: ClientTeamMetrics = [
-          { timestamp: 1000000, concurrentSandboxes: 10, sandboxStartRate: 5 },
-          { timestamp: 1100000, concurrentSandboxes: 20, sandboxStartRate: 10 },
-          { timestamp: 1205000, concurrentSandboxes: 30, sandboxStartRate: 15 }, // 105% of step
-        ]
-
-        // with 10% tolerance, 105% should be within tolerance
-        const resultLowTolerance = fillMetricsWithZeros(
-          data,
-          1000000,
-          1400000,
-          step,
-          0.1
-        )
-        const zerosLow = resultLowTolerance.filter(
-          (p) =>
-            p.timestamp > 1100000 &&
-            p.timestamp < 1205000 &&
-            p.concurrentSandboxes === 0
-        )
-        expect(zerosLow).toHaveLength(0) // no zeros, within tolerance
-
-        // with 3% tolerance, 105% should be anomalous
-        const resultHighTolerance = fillMetricsWithZeros(
-          data,
-          1000000,
-          1400000,
-          step,
-          0.03
-        )
-        const zerosHigh = resultHighTolerance.filter(
-          (p) =>
-            p.timestamp > 1100000 &&
-            p.timestamp < 1205000 &&
-            p.concurrentSandboxes === 0
-        )
-        expect(zerosHigh.length).toBeGreaterThan(0) // should add zeros
-      })
-    })
-  })
-
   describe('calculateAverage', () => {
     it('should calculate average of line chart data', () => {
       const data = [
